+++ asyncgit/src/push.rs
@@ -4,90 +4,15 @@ use crate::{
         cred::BasicAuthCredential, remotes::push::push,
         remotes::push::ProgressNotification,
     },
-    AsyncNotification, CWD,
+    AsyncNotification, RemoteProgress, CWD,
 };
 use crossbeam_channel::{unbounded, Receiver, Sender};
-use git2::PackBuilderStage;
 use std::{
-    cmp,
     sync::{Arc, Mutex},
     thread,
     time::Duration,
 };
 use thread::JoinHandle;
-
-///
-#[derive(Clone, Debug)]
-pub enum RemoteProgressState {
-    ///
-    PackingAddingObject,
-    ///
-    PackingDeltafiction,
-    ///
-    Pushing,
-    ///
-    Done,
-}
-
-///
-#[derive(Clone, Debug)]
-pub struct RemoteProgress {
-    ///
-    pub state: RemoteProgressState,
-    ///
-    pub progress: u8,
-}
-
-impl RemoteProgress {
-    ///
-    pub fn new(
-        state: RemoteProgressState,
-        current: usize,
-        total: usize,
-    ) -> Self {
-        let total = cmp::max(current, total) as f32;
-        let progress = current as f32 / total * 100.0;
-        let progress = progress as u8;
-        Self { state, progress }
-    }
-}
-
-impl From<ProgressNotification> for RemoteProgress {
-    fn from(progress: ProgressNotification) -> Self {
-        match progress {
-            ProgressNotification::Packing {
-                stage,
-                current,
-                total,
-            } => match stage {
-                PackBuilderStage::AddingObjects => {
-                    RemoteProgress::new(
-                        RemoteProgressState::PackingAddingObject,
-                        current,
-                        total,
-                    )
-                }
-                PackBuilderStage::Deltafication => {
-                    RemoteProgress::new(
-                        RemoteProgressState::PackingDeltafiction,
-                        current,
-                        total,
-                    )
-                }
-            },
-            ProgressNotification::PushTransfer {
-                current,
-                total,
-                ..
-            } => RemoteProgress::new(
-                RemoteProgressState::Pushing,
-                current,
-                total,
-            ),
-            _ => RemoteProgress::new(RemoteProgressState::Done, 1, 1),
-        }
-    }
-}
 
 ///
 #[derive(Default, Clone, Debug)]
@@ -293,6 +218,7 @@ impl AsyncPush {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::remote_progress::RemoteProgressState;
 
     #[test]
     fn test_progress_zero_total() {
